*** empty log message ***
git-svn-id: https://svn.library.emory.edu/svn/slavevoyages/trunk@582 2440b244-9809-4ec7-bebd-234906d9abf5

+++ src/edu/emory/library/tast/ui/search/query/searchables/SearchableAttributeSimpleNumeric.java
@@ -47,7 +47,7 @@ public class SearchableAttributeSimpleNumeric extends SearchableAttributeSimple
 		if (number == null) {
 			throw new NumberFormatException();
 		}
-		switch (((NumericAttribute)getAttributes()[0]).getType().intValue())
+		switch (((NumericAttribute)getAttributes()[0]).getType())
 		{
 		case NumericAttribute.TYPE_INTEGER:
 			return new Integer(number);
